bugfix: remove error in price with ProductCategoryDB schema

+++ app/entities/product.py
@@ -69,7 +69,7 @@ class ProductCategoryInDB(BaseModel):
     product_id: int
     name: str
     uri: str
-    price: int
+    price: Decimal
     active: bool
     direct_sales: bool
     description: Json | dict | None
